feat: user review and admin response review

+++ larose_be/src/main/java/com/larose/dto/BookingDTO.java
@@ -27,7 +27,7 @@ public class BookingDTO {
     private String roomTitle;
 
     // Thông tin loại phòng
-    private Short roomTypeId;
+    private Long roomTypeId;
     private String roomTypeName;
 
     // Thông tin user (nếu cần)
